backend/requirements.txt
# Django Backend Requirements
Django>=6.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0

# Async task processing
celery>=5.6.0
redis>=7.0.0

# Production server
gunicorn>=21.0.0
uvicorn>=0.27.0

# Utilities
python-dateutil>=2.8.0
